Fix Windows wheel build: c_long input_pair + DLL bundling
Two independent blockers that prevented the windows-latest matrix row from
producing a working wheel:

1. `update_and_1_out` took `input_pair: c_long`, but the deserialized kwarg is
   i64. c_long is i32 on Windows (LLP64), so passing the i64 was a hard E0308
   type error there (it only compiled on Linux/macOS, where c_long == i64).
   Take i64 and cast to c_long at the FFI call.

2. The maturin include glob `encomp/coolprop/libCoolProp.*` never matched
   MSVC's `CoolProp.dll` (no "lib" prefix) -- the loader already expects that
   name (win32 candidates) but the wheel never shipped it. Widen to *CoolProp.*

+++ rust/src/coolprop.rs
@@ -243,7 +243,7 @@ impl State<'_> {
     /// errors already surfaced from `state()`.
     pub fn update_and_1_out(
         &mut self,
-        input_pair: c_long,
+        input_pair: i64, // cast to c_long at the FFI call: c_long is i32 on Windows (LLP64)
         v1: &[f64],
         v2: &[f64],
         out_key: c_long,
@@ -262,7 +262,7 @@ impl State<'_> {
         unsafe {
             (self.cp.batch1)(
                 self.handle,
-                input_pair,
+                input_pair as c_long,
                 v1.as_ptr(),
                 v2.as_ptr(),
                 n as c_long,
